requirements.txt
streamlit>=1.28.0
ultralytics>=8.0.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
streamlit-lottie>=0.0.3
requests>=2.25.0
pandas>=1.3.0
tqdm>=4.62.0

